refactor: split off common
